BACKPORT: devfreq: add support to handle device suspend state
bwmon governor does not differentiate with device
suspend / resume state of the device, there will
be a requirement with some bw monitor device,
to not access it's register space when device is
in suspend state.

Avoid accessing bw monitor register space when
the device is in suspend state, add device suspend
state intelligence in bwmon governor.

Change-Id: Iba6e0475809b663fbd4c3d56fc086bf4c93c164d
Signed-off-by: Santosh Mardi <gsantosh@codeaurora.org>
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>